+++ RickAndMortyWiki/Controllers/MainViewController.swift
@@ -1,0 +1,64 @@
+//
+//  ViewController.swift
+//  RickAndMortyWiki
+//
+//  Created by Igor Fernandes on 02/11/22.
+//
+
+import UIKit
+
+class MainViewController: UIViewController {
+    private var allCharacters: [Results] = [Results]()
+    
+    lazy var mainView: MainView = {
+        let main = MainView()
+        main.collectionView.delegate = self
+        main.collectionView.dataSource = self
+        return main
+    }()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Welcome"
+        view.backgroundColor = .red
+        Service.getAllCharacters { [weak self] result in
+            switch result {
+            case .success(let success):
+                self?.allCharacters = success.results
+                print(self?.allCharacters[0].name)
+            case .failure(let failure):
+                print(failure)
+            }
+        }
+    }
+    
+    override func loadView() {
+        self.view = mainView
+    }
+}
+
+extension MainViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 10
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CharacterInfoCollectionViewCell.identifier, for: indexPath) as! CharacterInfoCollectionViewCell
+        cell.configure(with: Results(id: 1, name: "Morty", status: "Dead", species: "Human", type: "Unknown", gender: "Male", origin: Location(name: "", url: ""), location: Location(name: "", url: ""), image: "", episode: ["1"], url: "", created: ""))
+        cell.backgroundColor = .black
+        cell.layer.borderColor = UIColor.green.cgColor
+        cell.layer.borderWidth = 5
+        cell.layer.cornerRadius = 5
+        cell.clipsToBounds = true
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: self.view.bounds.width - 10, height: 200)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        collectionView.deselectItem(at: indexPath, animated: true)
+    }
+}
+
