+++ RickAndMortyWiki/Model/Character.swift
@@ -15,7 +15,7 @@ struct Results: Codable {
     let id: Int
     let name, status, species, type: String
     let gender: String
-    let origin, location: Location
+    let origin, location: Location?
     let image: String
     let episode: [String]
     let url: String
