+++ RickAndMortyWiki/View/Main/MainView.swift
@@ -20,8 +20,6 @@ class MainView: UIView {
         // create collection
         let collection = UICollectionView(frame: .zero, collectionViewLayout: layout)
         collection.backgroundColor = .purple
-        collection.delegate = self
-        collection.dataSource = self
         collection.translatesAutoresizingMaskIntoConstraints = false
         collection.register(CharacterInfoCollectionViewCell.self, forCellWithReuseIdentifier: CharacterInfoCollectionViewCell.identifier)
         return collection
@@ -42,27 +40,3 @@ class MainView: UIView {
     }
 }
 
-extension MainView: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 10
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CharacterInfoCollectionViewCell.identifier, for: indexPath)
-        cell.backgroundColor = .black
-        cell.layer.borderColor = UIColor.green.cgColor
-        cell.layer.borderWidth = 5
-        cell.layer.cornerRadius = 5
-        cell.clipsToBounds = true
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: self.bounds.width - 10, height: 200)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        collectionView.deselectItem(at: indexPath, animated: true)
-        didTapCharacter?()
-    }
-}
